feat: updating the integration test and setup to be done in before all
